requirements.txt
GitPython==3.1.45
gurobipy==12.0.3
hybrid==1.2.3
matplotlib==3.5.1
numpy==1.22.0
pandas==2.0.3
scipy==1.8.0
